+++ src/plugins/data_source_management/framework/catalog_cache/cache_intercept.ts
