+++ src/plugins/data_source_management/public/components/direct_query_data_sources_components/direct_query_acceleration_creation/create/create_acceleration.tsx
@@ -27,8 +27,8 @@ import {
   CreateAccelerationForm,
   DirectQueryLoadingStatus,
 } from '../../../../../framework/types';
-import { useLoadTableColumnsToCache } from '../../../../../framework/catlog_cache/cache_loader';
-import { CatalogCacheManager } from '../../../../../framework/catlog_cache/cache_manager';
+import { useLoadTableColumnsToCache } from '../../../../../framework/catalog_cache/cache_loader';
+import { CatalogCacheManager } from '../../../../../framework/catalog_cache/cache_manager';
 
 import { IndexAdvancedSettings } from '../selectors/index_advanced_settings';
 import { IndexSettingOptions } from '../selectors/index_setting_options';
