+++ src/plugins/data_source_management/public/components/direct_query_data_sources_components/direct_query_acceleration_creation/selectors/selector_helpers/load_databases.tsx
@@ -7,7 +7,7 @@ import { EuiButtonIcon, EuiLoadingSpinner } from '@elastic/eui';
 import React, { useEffect, useState } from 'react';
 import { HttpStart, NotificationsStart } from 'opensearch-dashboards/public';
 import { DirectQueryLoadingStatus } from '../../../../../../framework/types';
-import { useLoadDatabasesToCache } from '../../../../../../framework/catlog_cache/cache_loader';
+import { useLoadDatabasesToCache } from '../../../../../../framework/catalog_cache/cache_loader';
 
 interface SelectorLoadDatabasesProps {
   dataSourceName: string;
