+++ src/plugins/data_source_management/public/components/direct_query_data_sources_components/direct_query_acceleration_creation/selectors/selector_helpers/load_objects.test.tsx
@@ -13,9 +13,9 @@ import { DirectQueryLoadingStatus } from '../../../../../../framework/types';
 import {
   useLoadAccelerationsToCache,
   useLoadTablesToCache,
-} from '../../../../../../framework/catlog_cache/cache_loader';
+} from '../../../../../../framework/catalog_cache/cache_loader';
 
-jest.mock('../../../../../../framework/catlog_cache/cache_loader');
+jest.mock('../../../../../../framework/catalog_cache/cache_loader');
 
 const mockHttp: HttpStart = ({} as unknown) as HttpStart;
 const mockNotifications: NotificationsStart = ({
