+++ src/plugins/data_source_management/public/components/direct_query_data_sources_components/direct_query_acceleration_creation/selectors/selector_helpers/load_objects.tsx
@@ -10,7 +10,7 @@ import { DirectQueryLoadingStatus } from '../../../../../../framework/types';
 import {
   useLoadAccelerationsToCache,
   useLoadTablesToCache,
-} from '../../../../../../framework/catlog_cache/cache_loader';
+} from '../../../../../../framework/catalog_cache/cache_loader';
 
 interface SelectorLoadDatabasesProps {
   dataSourceName: string;
