+++ src/plugins/data_source_management/public/components/direct_query_data_sources_components/direct_query_acceleration_management/acceleration_details_flyout.test.tsx
@@ -13,7 +13,7 @@ import { CachedAcceleration } from '../../../../framework/types';
 import { useAccelerationOperation } from './acceleration_operation';
 
 jest.mock('./acceleration_operation');
-jest.mock('../../../../framework/catlog_cache/cache_loader');
+jest.mock('../../../../framework/catalog_cache/cache_loader');
 jest.mock('../../../../framework/utils/shared', () => ({
   DSL_BASE: '/api/accelerations',
 }));
