+++ src/plugins/data_source_management/public/components/direct_query_data_sources_components/direct_query_associated_object_management/associated_objects_details_flyout.tsx
@@ -33,8 +33,8 @@ import {
   CachedColumn,
   DirectQueryLoadingStatus,
 } from '../../../../framework/types';
-import { useLoadTableColumnsToCache } from '../../../../framework/catlog_cache/cache_loader';
-import { CatalogCacheManager } from '../../../../framework/catlog_cache/cache_manager';
+import { useLoadTableColumnsToCache } from '../../../../framework/catalog_cache/cache_loader';
+import { CatalogCacheManager } from '../../../../framework/catalog_cache/cache_manager';
 import {
   getRenderAccelerationDetailsFlyout,
   getRenderCreateAccelerationFlyout,
