+++ src/plugins/data_source_management/public/components/direct_query_data_sources_components/direct_query_associated_object_management/associated_objects_tab.tsx
@@ -27,7 +27,7 @@ import {
   DatasourceDetails,
   DirectQueryLoadingStatus,
 } from '../../../../framework/types';
-import { CatalogCacheManager } from '../../../../framework/catlog_cache/cache_manager';
+import { CatalogCacheManager } from '../../../../framework/catalog_cache/cache_manager';
 import { getRenderCreateAccelerationFlyout } from '../../../plugin';
 import {
   CreateAccelerationFlyoutButton,
