+++ src/plugins/data_source_management/public/components/frame_work_test/catlog_cache/cache_manager.test.ts
@@ -16,7 +16,7 @@ import {
   CachedDatabase,
   DataSourceCacheData,
 } from '../../../../framework/types';
-import { CatalogCacheManager } from '../../../../framework/catlog_cache/cache_manager';
+import { CatalogCacheManager } from '../../../../framework/catalog_cache/cache_manager';
 
 interface LooseObject {
   [key: string]: any;
